services/backend-services/pyproject.toml
[project]
name = "insurguard-backend"
version = "0.1.0"
description = "InsurGuard AI — FastAPI backend with Google Gemini and RAG pipeline"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=4.3.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "pillow>=10.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
